DRYed up NetCDF and Zarr sources.
