+++ intake_xarray/__init__.py
@@ -6,6 +6,7 @@
 # Miniconda installation.
 # -----------------------------------------------------------------------------
 from intake.source import base
+import xarray as xr
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
@@ -70,3 +71,41 @@ class ZarrPlugin(base.Plugin):
         base_kwargs, source_kwargs = self.separate_base_kwargs(kwargs)
         return ZarrSource(urlpath, storage_options, base_kwargs['metadata'],
                           **source_kwargs)
+
+
+class DataSourceMixin:
+
+    def _get_schema(self):
+        if self._ds is None:
+            self._open_dataset()
+
+        metadata = {
+            'dims': dict(self._ds.dims),
+            'data_vars': tuple(self._ds.data_vars.keys()),
+            'coords': tuple(self._ds.coords.keys())
+        }
+        metadata.update(self._ds.attrs)
+        return base.Schema(
+            datashape=None,
+            dtype=xr.Dataset,
+            shape=None,
+            npartitions=None,
+            extra_metadata=metadata)
+
+    def read(self):
+        self._load_metadata()
+        return self._ds.load()
+
+    def read_chunked(self):
+        self._load_metadata()
+        return self._ds
+
+    def read_partition(self, i):
+        raise NotImplementedError
+
+    def to_dask(self):
+        return self.read_chunked()
+
+    def close(self):
+        self._ds.close()
+        self._ds = None
