+++ intake_xarray/netcdf.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 import xarray as xr
 from intake.source import base
+from . import DataSourceMixin
 
-
-class NetCDFSource(base.DataSource):
+class NetCDFSource(DataSourceMixin, base.DataSource):
     """Open a xarray file.
 
     Parameters
@@ -27,42 +27,6 @@ class NetCDFSource(base.DataSource):
 
     def _open_dataset(self):
         url = self.urlpath
-        if "*" in url:
-            return xr.open_mfdataset(url, chunks=self.chunks, **self._kwargs)
-        else:
-            return xr.open_dataset(url, chunks=self.chunks, **self._kwargs)
+        _open_dataset = xr.open_mfdataset if "*" in url else xr.open_dataset
 
-    def _get_schema(self):
-        if self._ds is None:
-            self._ds = self._open_dataset()
-
-        metadata = {
-            'dims': dict(self._ds.dims),
-            'data_vars': tuple(self._ds.data_vars.keys()),
-            'coords': tuple(self._ds.coords.keys())
-        }
-        metadata.update(self._ds.attrs)
-        return base.Schema(
-            datashape=None,
-            dtype=xr.Dataset,
-            shape=None,
-            npartitions=None,
-            extra_metadata=metadata)
-
-    def read(self):
-        self._load_metadata()
-        return self._ds.load()
-
-    def read_chunked(self):
-        self._load_metadata()
-        return self._ds
-
-    def read_partition(self, i):
-        raise NotImplementedError
-
-    def to_dask(self):
-        return self.read_chunked()
-
-    def close(self):
-        self._ds.close()
-        self._ds = None
+        self._ds = _open_dataset(url, chunks=self.chunks, **self._kwargs)
